fix : fix tests

+++ src/main/java/nl/fontys/s3/huister/business/request/property/CreatePropertyRequest.java
@@ -15,5 +15,5 @@ public class CreatePropertyRequest {
     private int ownerId,cityId;
     private String streetName,postCode,description,cityName;
     private List<String> imageUrls;
-    private long area,price;
+    private double area,price;
 }
