+++ src/main/java/nl/fontys/s3/huister/business/response/property/GetAllPropertiesResponse.java
@@ -13,5 +13,5 @@ import lombok.NoArgsConstructor;
 public class GetAllPropertiesResponse {
     private int id;
     private String streetName,postCode,description,ownerName,cityName,imageUrl;
-    private long area,price;
+    private double area,price;
 }
